+++ tests/OrganizationJoinMfaValidationTest.php
@@ -1,18 +1,10 @@
 <?php
 namespace Keboola\ManageApiTest;
 
-use Keboola\ManageApi\Client;
 use Keboola\ManageApi\ClientException;
 
-class OrganizationJoinMfaValidationTest extends ClientTestCase
+class OrganizationJoinMfaValidationTest extends ClientMfaTestCase
 {
-    private const DUMMY_USER_EMAIL = 'spam+spam@keboola.com';
-
-    /** @var Client */
-    private $normalUserWithMfaClient;
-
-    private $normalUserWithMfa;
-
     private $organization;
 
     /**
@@ -23,13 +15,6 @@ class OrganizationJoinMfaValidationTest extends ClientTestCase
     public function setUp()
     {
         parent::setUp();
-
-        $this->normalUserWithMfaClient = new Client([
-            'token' => getenv('KBC_TEST_ADMIN_WITH_MFA_TOKEN'),
-            'url' => getenv('KBC_MANAGE_API_URL'),
-        ]);
-
-        $this->normalUserWithMfa = $this->normalUserWithMfaClient->verifyToken()['user'];
 
         $this->client->addUserToMaintainer($this->testMaintainerId, ['email' => self::DUMMY_USER_EMAIL]);
 
@@ -89,18 +74,5 @@ class OrganizationJoinMfaValidationTest extends ClientTestCase
 
         $member = $this->findOrganizationMember($this->organization['id'], $this->normalUser['email']);
         $this->assertNull($member);
-    }
-
-    protected function findOrganizationMember(int $organizationId, string $userEmail): ?array
-    {
-        $members = $this->normalUserWithMfaClient->listOrganizationUsers($organizationId);
-
-        foreach ($members as $member) {
-            if ($member['email'] === $userEmail) {
-                return $member;
-            }
-        }
-
-        return null;
     }
 }
