+++ tests/ProjectJoinRequestsMfaValidationTest.php
@@ -1,18 +1,10 @@
 <?php
 namespace Keboola\ManageApiTest;
 
-use Keboola\ManageApi\Client;
 use Keboola\ManageApi\ClientException;
 
-class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
+class ProjectJoinRequestsMfaValidationTest extends ClientMfaTestCase
 {
-    private const DUMMY_USER_EMAIL = 'spam+spam@keboola.com';
-
-    /** @var Client */
-    private $normalUserWithMfaClient;
-
-    private $normalUserWithMfa;
-
     private $organization;
 
     /**
@@ -25,13 +17,6 @@ class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
     public function setUp()
     {
         parent::setUp();
-
-        $this->normalUserWithMfaClient = new Client([
-            'token' => getenv('KBC_TEST_ADMIN_WITH_MFA_TOKEN'),
-            'url' => getenv('KBC_MANAGE_API_URL'),
-        ]);
-
-        $this->normalUserWithMfa = $this->normalUserWithMfaClient->verifyToken()['user'];
 
         $this->client->addUserToMaintainer($this->testMaintainerId, ['email' => self::DUMMY_USER_EMAIL]);
 
@@ -60,7 +45,7 @@ class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
 
     public function testSuperAdminWithoutMfaCannotRequestAccess()
     {
-        $projectId = $this->createProjectWithAdminHavingMfaEnabled();
+        $projectId = $this->createProjectWithAdminHavingMfaEnabled($this->organization['id']);
 
         $this->normalUserWithMfaClient->updateOrganization(
             $this->organization['id'],
@@ -87,7 +72,7 @@ class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
 
     public function testJoinRequestOfUserWithoutMfaCannotBeApproved()
     {
-        $projectId = $this->createProjectWithAdminHavingMfaEnabled();
+        $projectId = $this->createProjectWithAdminHavingMfaEnabled($this->organization['id']);
 
         $this->normalUserWithMfaClient->updateOrganization(
             $this->organization['id'],
@@ -130,7 +115,7 @@ class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
 
     public function testJoinRequestReject()
     {
-        $projectId = $this->createProjectWithAdminHavingMfaEnabled();
+        $projectId = $this->createProjectWithAdminHavingMfaEnabled($this->organization['id']);
 
         $this->normalUserWithMfaClient->updateOrganization(
             $this->organization['id'],
@@ -158,27 +143,5 @@ class ProjectJoinRequestsMfaValidationTest extends ClientTestCase
 
         $joinRequests = $this->client->listMyProjectJoinRequests();
         $this->assertCount(0, $joinRequests);
-    }
-
-    protected function findProjectUser(int $projectId, string $userEmail): ?array
-    {
-        $projectUsers = $this->normalUserWithMfaClient->listProjectUsers($projectId);
-
-        foreach ($projectUsers as $projectUser) {
-            if ($projectUser['email'] === $userEmail) {
-                return $projectUser;
-            }
-        }
-
-        return null;
-    }
-
-    private function createProjectWithAdminHavingMfaEnabled(): int
-    {
-        $project = $this->normalUserWithMfaClient->createProject($this->organization['id'], [
-            'name' => 'My test',
-        ]);
-
-        return $project['id'];
     }
 }
